requirements.txt
python-dotenv
google-generativeai
Pillow
streamlit
streamlit-mic-recorder